refactor: service 를 controller 로 수정
