+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.service.RacingCarGame;
+import racingcar.controller.RacingCarGame;
 
 public class Application {
     public static void main(String[] args) {
